Image SWAPS
Slack means something different to everyone.

If you have to ask, you can never know.
Slack is doing as little as possible, yet mysteriously benefiting at the same time.
It's a kind of direct perception, unfettered by 'so-called common sense.
Slack is not exactly laziness, but a kind of  'active sloth'.
True Slack is  "Something for Nothing"..

Everyone is born with Original Slack, but the Conspiracy has most of that now.